chore: fix gender picker, date picker, refactor require login view using High-Order component, change set token to store flow
