+++ components/profile/ProfileDatePicker.tsx
@@ -19,8 +19,8 @@ interface DatePickerInputProps {
 }
 
 const convertToDate = (dateStr: string) => {
-  const [month, day, year] = dateStr
-    .split("/")
+  const [year, month, day] = dateStr
+    .split("-")
     .map((part) => parseInt(part, 10))
   return new Date(year, month - 1, day)
 }
@@ -51,7 +51,7 @@ const ProfileDatePicker = ({
     const day = String(date.getDate()).padStart(2, "0")
     const month = String(date.getMonth() + 1).padStart(2, "0")
     const year = date.getFullYear()
-    return `${month}/${day}/${year}`
+    return `${year}-${month}-${day}`
   }
 
   const handleConfirm = (arg: any): void => {
